fix bugs on rendering the establishment data
