+++ app/user/book/index.tsx
@@ -6,9 +6,7 @@ import LinkComponent from "@/components/LinkComponent";
 const index = () => {
     return (
         <ResponsiveContainer>
-            <LinkComponent style={{ width: "auto", marginBottom: 16 }} href="../user">
-                ← Back to Dashboard
-            </LinkComponent>
+            <LinkComponent style={{ width: "auto", marginBottom: 16 }} href="../user" label="← Back to Dashboard" />
             <EstablishmentSearch guest={false} />
         </ResponsiveContainer>
     );
